+++ src/plugins/workspace/public/components/workspace_creator/index.tsx
@@ -4,4 +4,3 @@
  */
 
 export { WorkspaceCreator } from './workspace_creator';
-export { WorkspacePermissionSetting } from './workspace_permission_setting_panel';
