+++ src/plugins/workspace/public/components/workspace_form/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export { WorkspaceForm } from './workspace_form';
+export { WorkspaceFormSubmitData } from './types';
+export { WorkspaceOperationType } from './constants';
