+++ src/plugins/workspace/public/components/workspace_form/workspace_cancel_modal.tsx
