+++ src/plugins/workspace/public/components/workspace_form/workspace_icon_selector.tsx
