+++ src/plugins/workspace/public/components/workspace_form/workspace_permission_setting_panel.tsx
@@ -4,200 +4,19 @@
  */
 
 import React, { useCallback, useMemo, useState, useEffect } from 'react';
-import {
-  EuiFlexGroup,
-  EuiComboBox,
-  EuiFlexItem,
-  EuiButton,
-  EuiButtonIcon,
-  EuiButtonGroup,
-  EuiFormRow,
-  EuiText,
-  EuiSpacer,
-} from '@elastic/eui';
+import { EuiButton, EuiFormRow, EuiText, EuiSpacer } from '@elastic/eui';
 import { i18n } from '@osd/i18n';
-import { WorkspacePermissionMode } from '../../../common/constants';
-
-export enum WorkspacePermissionItemType {
-  User = 'user',
-  Group = 'group',
-}
-
-export type WorkspacePermissionSetting =
-  | { type: WorkspacePermissionItemType.User; userId: string; modes: WorkspacePermissionMode[] }
-  | { type: WorkspacePermissionItemType.Group; group: string; modes: WorkspacePermissionMode[] };
-
-enum PermissionModeId {
-  Read = 'read',
-  ReadAndWrite = 'read+write',
-  Admin = 'admin',
-}
-
-const permissionModeOptions = [
-  {
-    id: PermissionModeId.Read,
-    label: i18n.translate('workspace.form.permissionSettingPanel.permissionModeOptions.read', {
-      defaultMessage: 'Read',
-    }),
-  },
-  {
-    id: PermissionModeId.ReadAndWrite,
-    label: i18n.translate(
-      'workspace.form.permissionSettingPanel.permissionModeOptions.readAndWrite',
-      {
-        defaultMessage: 'Read & Write',
-      }
-    ),
-  },
-  {
-    id: PermissionModeId.Admin,
-    label: i18n.translate('workspace.form.permissionSettingPanel.permissionModeOptions.admin', {
-      defaultMessage: 'Admin',
-    }),
-  },
-];
-
-const optionIdToWorkspacePermissionModesMap: {
-  [key: string]: WorkspacePermissionMode[];
-} = {
-  [PermissionModeId.Read]: [WorkspacePermissionMode.LibraryRead, WorkspacePermissionMode.Read],
-  [PermissionModeId.ReadAndWrite]: [
-    WorkspacePermissionMode.LibraryWrite,
-    WorkspacePermissionMode.Read,
-  ],
-  [PermissionModeId.Admin]: [WorkspacePermissionMode.LibraryWrite, WorkspacePermissionMode.Write],
-};
-
-const generateWorkspacePermissionItemKey = (
-  item: Partial<WorkspacePermissionSetting>,
-  index?: number
-) =>
-  [
-    ...(item.type ?? []),
-    ...(item.type === WorkspacePermissionItemType.User ? [item.userId] : []),
-    ...(item.type === WorkspacePermissionItemType.Group ? [item.group] : []),
-    ...(item.modes ?? []),
-    index,
-  ].join('-');
-
-// default permission mode is read
-const getPermissionModeId = (modes: WorkspacePermissionMode[]) => {
-  for (const key in optionIdToWorkspacePermissionModesMap) {
-    if (optionIdToWorkspacePermissionModesMap[key].every((mode) => modes?.includes(mode))) {
-      return key;
-    }
-  }
-  return PermissionModeId.Read;
-};
-
-interface WorkspacePermissionSettingInputProps {
-  index: number;
-  deletable: boolean;
-  type: WorkspacePermissionItemType;
-  userId?: string;
-  group?: string;
-  modes?: WorkspacePermissionMode[];
-  onGroupOrUserIdChange: (
-    groupOrUserId:
-      | { type: WorkspacePermissionItemType.User; userId?: string }
-      | { type: WorkspacePermissionItemType.Group; group?: string },
-    index: number
-  ) => void;
-  onPermissionModesChange: (
-    WorkspacePermissionMode: WorkspacePermissionMode[],
-    index: number
-  ) => void;
-  onDelete: (index: number) => void;
-}
-
-const WorkspacePermissionSettingInput = ({
-  index,
-  type,
-  userId,
-  group,
-  modes,
-  deletable,
-  onDelete,
-  onGroupOrUserIdChange,
-  onPermissionModesChange,
-}: WorkspacePermissionSettingInputProps) => {
-  const groupOrUserIdSelectedOptions = useMemo(
-    () => (group || userId ? [{ label: (group || userId) as string }] : []),
-    [group, userId]
-  );
-
-  const permissionModesSelectedId = useMemo(() => getPermissionModeId(modes ?? []), [modes]);
-  const handleGroupOrUserIdCreate = useCallback(
-    (groupOrUserId) => {
-      onGroupOrUserIdChange(
-        type === WorkspacePermissionItemType.Group
-          ? { type, group: groupOrUserId }
-          : { type, userId: groupOrUserId },
-        index
-      );
-    },
-    [index, type, onGroupOrUserIdChange]
-  );
-
-  const handleGroupOrUserIdChange = useCallback(
-    (options) => {
-      if (options.length === 0) {
-        onGroupOrUserIdChange({ type }, index);
-      }
-    },
-    [index, type, onGroupOrUserIdChange]
-  );
-
-  const handlePermissionModeOptionChange = useCallback(
-    (id: string) => {
-      if (optionIdToWorkspacePermissionModesMap[id]) {
-        onPermissionModesChange([...optionIdToWorkspacePermissionModesMap[id]], index);
-      }
-    },
-    [index, onPermissionModesChange]
-  );
-
-  const handleDelete = useCallback(() => {
-    onDelete(index);
-  }, [index, onDelete]);
-
-  return (
-    <EuiFlexGroup alignItems="center" gutterSize="l">
-      <EuiFlexItem>
-        <EuiComboBox
-          singleSelection
-          selectedOptions={groupOrUserIdSelectedOptions}
-          onCreateOption={handleGroupOrUserIdCreate}
-          onChange={handleGroupOrUserIdChange}
-          placeholder="Select"
-          style={{ width: 200 }}
-          data-test-subj={`workspaceForm-permissionSettingPanel-${index}-userId`}
-        />
-      </EuiFlexItem>
-      <EuiFlexItem grow={false}>
-        <EuiButtonGroup
-          type="single"
-          isDisabled={!deletable}
-          legend="Permission Modes"
-          options={permissionModeOptions}
-          idSelected={permissionModesSelectedId}
-          onChange={handlePermissionModeOptionChange}
-        />
-      </EuiFlexItem>
-      <EuiFlexItem grow={false}>
-        <EuiButtonIcon
-          color="danger"
-          aria-label="Delete permission setting"
-          iconType="trash"
-          display="base"
-          size="m"
-          onClick={handleDelete}
-          isDisabled={!deletable}
-        />
-      </EuiFlexItem>
-    </EuiFlexGroup>
-  );
-};
+import { WorkspacePermissionSetting } from './types';
+import {
+  WorkspacePermissionItemType,
+  optionIdToWorkspacePermissionModesMap,
+  PermissionModeId,
+} from './constants';
+import {
+  WorkspacePermissionSettingInput,
+  WorkspacePermissionSettingInputProps,
+} from './workspace_permission_setting_input';
+import { generateWorkspacePermissionItemKey, getPermissionModeId } from './utils';
 
 interface WorkspacePermissionSettingPanelProps {
   errors?: string[];
